refactor: utils/types 타입 관리 파일 생성 후 리팩토링
