+++ client/src/components/common/Gnb.tsx
@@ -1,6 +1,8 @@
 import React, { Dispatch } from 'react';
 import styled, { css } from 'styled-components';
 import { Link } from 'react-router-dom';
+import { useRecoilValue, useRecoilState } from 'recoil';
+import { modalVisibleStates, rightModalStates } from 'recoil/modal';
 import { ReactComponent as GnbHome } from 'images/gnb-home.svg';
 import { ReactComponent as GnbGroup } from 'images/gnb-group.svg';
 import { ReactComponent as GnbHomeActive } from 'images/gnb-home-active.svg';
@@ -8,34 +10,9 @@ import { ReactComponent as GnbGroupActive } from 'images/gnb-group-active.svg';
 import gnbMessage from 'images/gnb-message.svg';
 import gnbAlarm from 'images/gnb-alarm.svg';
 import gnbSelector from 'images/gnb-down-arrow.svg';
-import { UserSearchBar, UserSearchModal } from '..';
-import { useRecoilValue, useRecoilState } from 'recoil';
-import { modalVisibleStates, rightModalStates } from 'recoil/modal';
 import profileDefault from 'images/profile-default.png';
-
-type GnbProps = {
-  type?: string;
-};
-
-type FlexProps = {
-  center?: boolean;
-};
-
-type TabProps = {
-  current?: boolean;
-};
-
-type IconProps = {
-  img: any;
-};
-
-type RightModalProps = {
-  [key: string]: boolean;
-  rightModalFlag: boolean;
-  messageFlag: boolean;
-  alarmFlag: boolean;
-  selectorFlag: boolean;
-};
+import { UserSearchBar, UserSearchModal } from '..';
+import { GnbProps, FlexProps, TabProps, IconProps, RightModalProps } from 'utils/types';
 
 const GnbContainer = styled.div`
   width: 100%;
